jsonld.canonize workers implementation
